refactor useEffects for setPlayer and highlightMovements
