+++ src/Components/Board.tsx
@@ -1,4 +1,4 @@
-import { useRef, ReactNode } from 'react';
+import { useEffect, useRef, ReactNode } from 'react';
 import Tiles from './Tiles';
 import NewTileArea from './NewTileArea';
 import Pawns from './Pieces/Pawns';
@@ -10,6 +10,10 @@ import useHighlightArea from '../utils/useHighlightArea';
 import Timer from './Timer';
 import Pinged from './Pinged';
 import { useGameStartedDocState } from '../Contexts/FirestoreContext';
+import { usePlayerState } from '../Contexts/PlayerContext';
+import { usePlayerDocState } from '../Contexts/FirestoreContext';
+import { getDoc } from '../utils/useFirestore';
+import { Room } from '../types';
 
 const BoardComponent = ({timer, pinged, children}: {timer: ReactNode, pinged: ReactNode, children: ReactNode}) => {
   console.log('*** Board Component re render')
@@ -17,6 +21,23 @@ const BoardComponent = ({timer, pinged, children}: {timer: ReactNode, pinged: Re
   const { gameState } = useGame();
   const gameStarted = useGameStartedDocState();
   const [availableArea, highlightNewTileArea, clearHighlightAreas] = useHighlightArea(gameState.roomId);
+
+  const playerState = usePlayerState();
+  const {setPlayer} = usePlayerDocState();
+
+  useEffect(() => {
+    (async () => {
+      if (!gameStarted) return;
+
+      const docSnap = await getDoc(gameState.roomId);
+      if (!docSnap.exists()) return;
+      const roomFound = docSnap.data() as Room;
+
+      const currentPlayer = roomFound.players.find((player) => player.number === playerState.number);
+      if (!currentPlayer) return;
+      setPlayer(currentPlayer);
+    })()
+  }, [gameStarted])
 
   return (
     <>
