+++ src/Components/Pieces/Pawn.tsx
@@ -1,7 +1,8 @@
+import { useEffect } from 'react';
 import { Room, DBHeroPawn, DBTile } from '../../types';
 import { tileWallSize } from '../../constants';
 import { useGame } from '../../Contexts/GameContext';
-import { getDoc } from '../../utils/useFirestore';
+import { getDoc, setDoc } from '../../utils/useFirestore';
 import { showMovableSpaces } from '../../Helpers/TileMethods';
 import { 
   useTilesDocState,
@@ -9,6 +10,7 @@ import {
   useGamePausedDocState,
 } from '../../Contexts/FirestoreContext';
 import { getDisplacementValue } from '../../Helpers/TileMethods';
+import { getPlayerPawnActions } from '../../Helpers/PawnMethods';
 
 interface pawnProps {
   pawnData: DBHeroPawn,
@@ -22,6 +24,37 @@ const Pawn = ({pawnData}: pawnProps) => {
 
   const { player } = usePlayerDocState();
   const tiles: DBTile[] = useTilesDocState();
+
+  // TODO: HOC on a pawn component, modify logic for single pawn re render
+  useEffect(() => {
+    (async () => {
+      if (tiles.length <= 1) return;
+
+      const docSnap = await getDoc(gameState.roomId);
+      if (!docSnap.exists()) return;
+
+      const roomFound: Room = docSnap.data() as Room;
+      const newPawns = roomFound.pawns;
+
+      if (!player) return;
+      if (pawnData.playerHeld === player.number) {
+        const playerPawnActions = getPlayerPawnActions(player, tiles, newPawns, pawnData);
+          newPawns[color].blockedPositions = playerPawnActions.blockedPositions
+          newPawns[color].showMovableDirections = playerPawnActions.showMovableDirections
+          newPawns[color].showEscalatorSpaces = playerPawnActions.showEscalatorSpaces
+          newPawns[color].showTeleportSpaces = playerPawnActions.showTeleportSpaces
+      }
+
+      await setDoc(
+        gameState.roomId, 
+        { 
+          pawns: {
+            ...newPawns,
+          }
+        },
+      )
+    })()
+  }, [tiles.length])
 
   const toggleMovableSpaces = async () => {
     const docSnap = await getDoc(gameState.roomId);
