+++ src/Components/Pieces/Pawns.tsx
@@ -14,43 +14,43 @@ import { getPlayerPawnActions } from '../../Helpers/PawnMethods';
 import Pawn from './Pawn';
 
 const Pawns = () => {
-  const { gameState } = useGame();
-  const { player } = usePlayerDocState();
-  const tiles: DBTile[] = useTilesDocState();
+  // const { gameState } = useGame();
+  // const { player } = usePlayerDocState();
+  // const tiles: DBTile[] = useTilesDocState();
 
-  useEffect(() => {
-    (async () => {
-      if (tiles.length <= 1) return;
+  // useEffect(() => {
+  //   (async () => {
+  //     if (tiles.length <= 1) return;
 
-      const docSnap = await getDoc(gameState.roomId);
-      if (!docSnap.exists()) return;
+  //     const docSnap = await getDoc(gameState.roomId);
+  //     if (!docSnap.exists()) return;
 
-      const roomFound: Room = docSnap.data() as Room;
-      const newPawns = roomFound.pawns;
+  //     const roomFound: Room = docSnap.data() as Room;
+  //     const newPawns = roomFound.pawns;
 
-      Object.values(newPawns).forEach((pawn: DBHeroPawn) => {
-        const playerPawnActions = getPlayerPawnActions(player, tiles, newPawns, pawn);
-        if (pawn.playerHeld === player.number) {
-          pawn.blockedPositions = playerPawnActions.blockedPositions
-          pawn.showMovableDirections = playerPawnActions.showMovableDirections
-          pawn.showEscalatorSpaces = playerPawnActions.showEscalatorSpaces
-          pawn.showTeleportSpaces = playerPawnActions.showTeleportSpaces
-        }
-        console.log('newPawns', {playerPawnActions, pawn, blockedPositions: playerPawnActions.blockedPositions})
-      })
+  //     Object.values(newPawns).forEach((pawn: DBHeroPawn) => {
+  //       const playerPawnActions = getPlayerPawnActions(player, tiles, newPawns, pawn);
+  //       if (pawn.playerHeld === player.number) {
+  //         pawn.blockedPositions = playerPawnActions.blockedPositions
+  //         pawn.showMovableDirections = playerPawnActions.showMovableDirections
+  //         pawn.showEscalatorSpaces = playerPawnActions.showEscalatorSpaces
+  //         pawn.showTeleportSpaces = playerPawnActions.showTeleportSpaces
+  //       }
+  //       console.log('newPawns', {playerPawnActions, pawn, blockedPositions: playerPawnActions.blockedPositions})
+  //     })
 
-      console.log('newPawns', newPawns)
+  //     console.log('newPawns', newPawns)
 
-      await setDoc(
-        gameState.roomId, 
-        { 
-          pawns: {
-            ...newPawns,
-          }
-        },
-      )
-    })()
-  }, [tiles])
+  //     await setDoc(
+  //       gameState.roomId, 
+  //       { 
+  //         pawns: {
+  //           ...newPawns,
+  //         }
+  //       },
+  //     )
+  //   })()
+  // }, [tiles])
 
   return (
     <>
